Refactor accounts tests into separate files
